Turn on -Wextra -Wno-sign-compare -Wno-unused-paramter and fix issues.
We turn on a few extra warnings and fix the fallout that occurs
when building with --enable-developer.  Note that we get different
warnings on different machines and so this will be a work in
progress.  So far, we have built on NetBSD/amd64 5.99.64 (which
uses gcc 4.5.3) and Ubuntu 10.04.3 LTS (which uses gcc 4.4.3).

+++ base/test_base.c
@@ -279,7 +279,7 @@ test_json(void)
 	for (k = strlen(j[i]) - 1; k > 0; k--) {
 	    o = heim_json_create_with_bytes(j[i], k, 10, 0, NULL);
 	    if (o != NULL) {
-		fprintf(stderr, "Invalid JSON parsed: %.*s\n", k, j[i]);
+		fprintf(stderr, "Invalid JSON parsed: %.*s\n", (int)k, j[i]);
 		return EINVAL;
 	    }
 	}
